Add first part of liberator.
It seems like overkill at this point, but it will allow for great
flexibility in the future. For instance, it could be extended to
use magiccards.info as a source. Or it could support rate limiting,
memoization, etc.

I'm sure using a library like Scrapy would be better, but that might
be too much overkill. Or I just suffer from not invented here
snydrome.