Update Card model.
I decided to remove the set and rarity information from the card
model. That will go in another model, probably PrintedCard or
something. That way to information won't be duplicated, since
combinations of CardAtoms are always consistent across printings.

Also added a "type" attribute to the Card model. It describes what
kind of card it is: normal, double-faced, flip, or split.