Update the scraper.
It does *everything* now. First it gets a set's checklist, for all
the printing details. Then it gets the set's spoiler, for all the
card details. Then it gets each card's page, for flavor text and
rulings. Finally it gets the card's image.

A couple things still trip it up. It can't handle non-Oracle cards,
like the Un- sets. It also has trouble with double-faced, flip, and
split cards.