Add collection list and detail views.
I'm not sure about the URL scheme. I copied everything from the
decks, but I'm not sure it makes sense. Collections map one-to-one
with users and they conceptually belong to users. So it might make
more sense to use "/users/taylor/collection/", for example. That
would require getting rid of the top-level "/collections/" URL,
which isn't very useful in the first place.